compile errors still

+++ lxd/device/unix_hotplug.go
@@ -151,11 +151,11 @@ func (d *unixHotplug) Start() (*RunConfig, error) {
 		}
 	}
 
-	major, err := strconv.ParseUint(device.SysattrValue("MAJOR"), 10, 32)
+	major, err := uint32(strconv.ParseUint(device.SysattrValue("MAJOR"), 10, 32))
 	if err != nil {
 		return nil, err
 	}
-	minor, err := strconv.ParseUint(device.SysattrValue("MINOR"), 10, 32)
+	minor, err := uint32(strconv.ParseUint(device.SysattrValue("MINOR"), 10, 32))
 	if err != nil {
 		return nil, err
 	}
